Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,6 @@
 # include <stdio.h>
-float main(){
-    int a;
+int main(){
+    float a;
     scanf("%f",&a);
     printf("You entered: %.2f",a);
     return 0;
